Source Bing Ads: refactor
